+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # HEAD:
 
 - Fixed: false positives in inline comment detecting by `findCommentsInRaws` if a comment is the first/last in a file/line
+- Fixed: `findCommentsInRaws` error in function detection
 
 # 1.4.1